+++ Orchard-Swift/iOS/iPods.swift
@@ -40,18 +40,6 @@ public enum iPods {
         self = .unknown
 
         switch model {
-//        case "iPod1,1":
-//            self = .iPodTouch
-//
-//        case "iPod2,1":
-//            self = .iPodTouch2
-//
-//        case "iPod3,1":
-//            self = .iPodTouch3
-//
-//        case "iPod4,1":
-//            self = .iPodTouch4
-//
         case "iPod5,1":
             self = .iPodTouch5