+++ Package.swift
@@ -6,7 +6,7 @@ import PackageDescription
 let package = Package(
     name: "Orchard",
     platforms: [
-        .iOS(.v8),
+        .iOS(.v9),
         .watchOS(.v2),
         .tvOS(.v9)
     ],